refactor: extract all ffi specific code from flipt-evaluation
