+++ flipt-engine-ffi/Cargo.toml
@@ -11,6 +11,10 @@ publish = false
 libc = "0.2.150"
 serde = { version = "1.0.147", features = ["derive"] }
 serde_json = { version = "1.0.89", features = ["raw_value"] }
+reqwest = { version = "0.11.22", features = ["json", "blocking"] }
+tokio = { version = "1.33.0", features = ["full"] }
+futures = "0.3"
+openssl = { version = "0.10", features = ["vendored"] }
 
 [dependencies.flipt-evaluation]
 path = "../flipt-evaluation"
