+++ flipt-engine-wasm/src/lib.rs
@@ -1,3 +1,1 @@
-pub struct Engine {
-    
-}
+pub struct Engine {}
