+++ flipt-evaluation/Cargo.toml
@@ -12,16 +12,12 @@ name = "fliptevaluation"
 path = "src/lib.rs"
 
 [dependencies]
-reqwest = { version = "0.11.22", features = ["json", "blocking"] }
-tokio = { version = "1.33.0", features = ["full"] }
 serde = { version = "1.0.147", features = ["derive"] }
 serde_json = { version = "1.0.89", features = ["raw_value"] }
 uuid = { version = "1.5.0", features = ["v4"] }
 crc32fast = "1.3.2"
-chrono = { version = "0.4.31", features = ["serde", "rustc-serialize"] }
-futures = "0.3"
 thiserror = "1.0.50"
-openssl = { version = "0.10", features = ["vendored"] }
+chrono = { version = "0.4.31", features = ["serde", "rustc-serialize"] }
 
 [dev-dependencies]
 mockall = "0.12.1"
